Move Feature_Collection inside Feature_Registry file
